requirements.txt
colorama>=0.4.4
tqdm>=4.62.0
pyyaml>=6.0
jinja2>=3.0.0
gitignore-parser>=0.0.8
gitpython>=3.1.30 